requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
CacheControl==0.14.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
comtypes==1.4.9
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
dask==2024.12.1
datasets==3.1.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
emoji==2.13.2
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
firebase-admin==6.5.0
flatbuffers==24.3.25
fonttools==4.54.1
foundry==0.0.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
graphviz==0.20.3
grpcio==1.67.0
grpcio-status==1.67.0
gTTS==2.5.3
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.5.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazypredict==0.2.12
Levenshtein==0.26.0
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.43.0
locket==1.0.0
LughaatNLP==1.0.6
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.8
narwhals==1.6.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
openai-whisper==20240930
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.12.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==0.5.6
peft==0.13.2
pillow==10.4.0
platformdirs==4.3.2
plotly==5.24.1
preshed==3.0.9
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-Levenshtein==0.26.0
pyttsx3==2.98
pytz==2024.1
PyWavelets==1.7.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rouge_score==0.1.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
slither==0.3.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
solidity-parser==0.1.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
sqlparse==0.5.2
srsly==2.4.8
stack-data==0.6.3
stanza==1.9.2
statsmodels==0.14.2
streamlit==1.38.0
sympy==1.13.3
ta-lib @ file:///C:/Users/alij7/Downloads/ta_lib-0.6.0-cp312-cp312-win_amd64.whl#sha256=ba76393c412deeaead700acda6c4a4a4805195ec7715bbb88f4691599cab2752
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==3.2.1
tensorflow-metadata==1.16.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.18.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.20.4
toml==0.10.2
toolz==1.0.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.3
trl==0.12.1
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
urduhack==1.0.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xgboost==2.1.1
xxhash==3.5.0
yarl==1.18.0
